requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
pydantic==1.10.7
openai>=1.0.0
pydub==0.25.1
python-dotenv==1.0.0
requests==2.31.0
loguru==0.7.2
pypdf==3.17.1
tenacity==8.2.3
stripe
httpx
# Make these optional to handle cases where ffmpeg is not available
numpy>=1.20.0
scipy>=1.6.0
